Update loop cli options and rename Window Policy
To make it much clearer what it does - which is to automatically
set a view when looping we call it Loop view instead.

This changes both cli options (new --loop-view instead of --window-policy)
as well as internal structs.

We also change --idle-timeout to --loop-timeout so that all loop options
start with loop.

This also adds some short options: -l for --loop and -t for --loop-tieout.