Remove clause AND operator
Got added pretty much by default since if you have an OR, you obviously
need an AND. Well, looking at examples was not that useful, the primary
reason is that inter-op behavior behaves like AND already.

It would be useless except for the fact that clauses would commit
even if later AND-connected clauses failed, so it was not always
equivalent to stripping the AND. However you could get the same output
by either stripping the and or converting to THEN depending on the
failure mode.

It is not obvious why this would ever be needed and so away it goes!