Delete 'goto' op -> add 'skip to' op
There have been confusion about both the behavior of goto
(thinking it controls the next op executed) and that there is a skip to
(which has never existed).

Since a core tenet of fiskta is to be intuitive - we can change things
to how they probably should have been in all along. Skip already moves
the cursor, so 'skip to' makes more sense that a separate goto op.