Remove Makefile add simple build script
We'll be using Zig to provide release builds for all platforms, but
we want users who don't have or want Zig to be able to build easily
without having to figure out the command. A Makefile is overkill for
this purpose. Let's just have a very simple bash script that compiles
our single binary.